Modify MissionsTable.js file
